+++ NEWS.md
@@ -1,5 +1,9 @@
 # *News*
 
+# zipcodeR 0.3.3
+- This update vectorizes the `zip_distance()` function to allow distance calculations between two vectors or columns of ZIP codes. The function now returns a data.frame of the resulting distance calculation.
+- `zip_distance()` now includes an additional argument, units, which allows selection between miles and meters for distance calculations.
+
 # zipcodeR 0.3.2
 - `zip_code_db` has been updated.
 - `download_zip_data()` has been refactored to make data updates more easily accessible and compare against existing data. Data is now directly downloaded from the upstream source and the existing data GitHub repository will no longer be updated.